+++ legacy/src/handler/handler_falcon_sign.c
@@ -42,11 +42,11 @@
 #include <stdbool.h>
 #include <stddef.h>
 #include <string.h>
-#include "../sw.h"
+#include "sw.h"
 #include "os.h"
 #include "cx.h"
 #include "buffer.h"
-#include "../globals.h"
+#include "globals.h"
 #include "send_response.h"
 #include "zkn_errors.h"
 #include "handler_falcon_sign.h"