+++ tools/sim/shim/bip32.h
@@ -1,0 +1,2 @@
+#pragma once
+#define MAX_BIP32_PATH 10