+++ tools/sim/shim/macros.h
@@ -1,0 +1,2 @@
+#pragma once
+#define MEMBER_SIZE(type, member) (sizeof(((type *)0)->member))