+++ tools/sim/shim/status_words.h
@@ -1,0 +1,8 @@
+#pragma once
+#define SWO_SUCCESS                  0x9000
+#define SWO_INCORRECT_DATA           0x6A80
+#define SWO_INCORRECT_P1_P2          0x6A86
+#define SWO_WRONG_DATA_LENGTH        0x6A87
+#define SWO_INVALID_CLA              0x6E00
+#define SWO_INVALID_INS              0x6D00
+#define SWO_CONDITIONS_NOT_SATISFIED 0x6985